types: changed flags back to 8-bits

+++ src/md_types.h
@@ -40,7 +40,7 @@ typedef int32_t     md_residue_id_t;
 typedef int32_t     md_chain_idx_t;
 typedef int32_t     md_bond_idx_t;
 typedef uint32_t    md_secondary_structure_t;
-typedef uint32_t    md_flags_t;
+typedef uint8_t     md_flags_t;
 typedef uint8_t     md_element_t;
 typedef uint8_t     md_ramachandran_type_t;
 typedef uint8_t     md_valence_t;
